+++ rust/optify/src/builder/mod.rs
@@ -1,7 +1,7 @@
 pub(crate) mod builder_impl;
 pub(crate) mod builder_trait;
-pub(crate) mod watchable_builder;
+pub(crate) mod watcher_builder;
 
 pub use builder_impl::*;
 pub use builder_trait::*;
-pub use watchable_builder::*;
+pub use watcher_builder::*;
