+++ rust/optify/src/provider/mod.rs
@@ -1,4 +1,7 @@
 pub(crate) mod provider_impl;
 pub(crate) mod provider_trait;
+pub(crate) mod watcher;
+
 pub use provider_impl::*;
 pub use provider_trait::*;
+pub use watcher::*;
